Change project struct
