--- ArciverExecutor.h
@@ -1,12 +1,0 @@
-#ifndef KURSOVAYA_EXECUTEFANO_H
-#define KURSOVAYA_EXECUTEFANO_H
-
-#include <string>
-#include "Config.h"
-
-class ArchiverExecutor{
-public:
-    static void execute(Config &config);
-};
-
-#endif //KURSOVAYA_EXECUTEFANO_H
