+++ CMakeLists.txt
@@ -3,4 +3,4 @@ project(CourseWork)
 
 set(CMAKE_CXX_STANDARD 23)
 
-add_executable(CourseWork main.cpp ArciverExecutor.h ArciverExecutor.cpp Fano.h Fano.cpp tree.cpp tree.h SyntaxChecker.cpp SyntaxChecker.h Exception.cpp Exception.h Config.h ConfigMapper.cpp ConfigMapper.h Archive.cpp Archive.h Log.cpp Log.h Utils.cpp Utils.h)
+add_executable(CourseWork main.cpp ArchiverExecutor.h ArchiverExecutor.cpp Fano.h Fano.cpp tree.cpp tree.h utils/Exception.cpp utils/Exception.h utils/ConfigMapper.cpp utils/ConfigMapper.h Archive.cpp Archive.h utils/Log.cpp utils/Log.h utils/Utils.cpp utils/Utils.h)
