+++ cmake-build-debug/.cmake/api/v1/reply/target-CourseWork-Debug-85e3f3c79f2820a71243.json
@@ -60,11 +60,10 @@
 				4,
 				5,
 				7,
-				9,
+				10,
 				12,
 				14,
-				16,
-				18
+				16
 			]
 		}
 	],
@@ -106,11 +105,10 @@
 				4,
 				5,
 				7,
-				9,
+				10,
 				12,
 				14,
-				16,
-				18
+				16
 			]
 		},
 		{
@@ -121,12 +119,11 @@
 				3,
 				6,
 				8,
-				10,
+				9,
 				11,
 				13,
 				15,
-				17,
-				19
+				17
 			]
 		}
 	],
@@ -169,17 +166,6 @@
 		{
 			"backtrace" : 1,
 			"path" : "tree.h",
-			"sourceGroupIndex" : 1
-		},
-		{
-			"backtrace" : 1,
-			"compileGroupIndex" : 0,
-			"path" : "SyntaxChecker.cpp",
-			"sourceGroupIndex" : 0
-		},
-		{
-			"backtrace" : 1,
-			"path" : "SyntaxChecker.h",
 			"sourceGroupIndex" : 1
 		},
 		{
