+++ cmake-build-debug/CMakeFiles/CourseWork.dir/ArciverExecutor.cpp.obj.d
@@ -120,7 +120,7 @@ CMakeFiles/CourseWork.dir/ArciverExecutor.cpp.obj: \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/cerrno \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/charconv.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/basic_string.tcc \
- C:\Users\NHTHS\Desktop\CourseWork\Config.h \
+ C:\Users\NHTHS\Desktop\CourseWork\ConfigMapper.h \
  C:\Users\NHTHS\Desktop\CourseWork\Log.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/iostream \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/ostream \
