+++ cmake-build-debug/CMakeFiles/CourseWork.dir/ConfigMapper.cpp.obj.d
@@ -1,7 +1,7 @@
 CMakeFiles/CourseWork.dir/ConfigMapper.cpp.obj: \
  C:\Users\NHTHS\Desktop\CourseWork\ConfigMapper.cpp \
  C:\Users\NHTHS\Desktop\CourseWork\ConfigMapper.h \
- C:\Users\NHTHS\Desktop\CourseWork\Config.h \
+ C:\Users\NHTHS\Desktop\CourseWork\Log.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/string \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/x86_64-w64-mingw32/bits/c++config.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/x86_64-w64-mingw32/bits/os_defines.h \
@@ -121,6 +121,4 @@ CMakeFiles/CourseWork.dir/ConfigMapper.cpp.obj: \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/cerrno \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/charconv.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/basic_string.tcc \
- C:\Users\NHTHS\Desktop\CourseWork\Log.h \
- C:\Users\NHTHS\Desktop\CourseWork\SyntaxChecker.h \
  C:\Users\NHTHS\Desktop\CourseWork\Exception.h
