+++ cmake-build-debug/CMakeFiles/CourseWork.dir/DependInfo.cmake
@@ -14,7 +14,6 @@ set(CMAKE_DEPENDS_DEPENDENCY_FILES
   "C:/Users/NHTHS/Desktop/CourseWork/Exception.cpp" "CMakeFiles/CourseWork.dir/Exception.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/Exception.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/Fano.cpp" "CMakeFiles/CourseWork.dir/Fano.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/Fano.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/Log.cpp" "CMakeFiles/CourseWork.dir/Log.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/Log.cpp.obj.d"
-  "C:/Users/NHTHS/Desktop/CourseWork/SyntaxChecker.cpp" "CMakeFiles/CourseWork.dir/SyntaxChecker.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/SyntaxChecker.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/Utils.cpp" "CMakeFiles/CourseWork.dir/Utils.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/Utils.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/main.cpp" "CMakeFiles/CourseWork.dir/main.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/main.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/tree.cpp" "CMakeFiles/CourseWork.dir/tree.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/tree.cpp.obj.d"
