+++ cmake-build-debug/CMakeFiles/CourseWork.dir/main.cpp.obj.d
@@ -120,7 +120,7 @@ CMakeFiles/CourseWork.dir/main.cpp.obj: \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/cerrno \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/charconv.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/basic_string.tcc \
- C:\Users\NHTHS\Desktop\CourseWork\Config.h \
+ C:\Users\NHTHS\Desktop\CourseWork\ConfigMapper.h \
  C:\Users\NHTHS\Desktop\CourseWork\Log.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/iostream \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/ostream \
@@ -150,6 +150,4 @@ CMakeFiles/CourseWork.dir/main.cpp.obj: \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/ostream.tcc \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/istream \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/istream.tcc \
- C:\Users\NHTHS\Desktop\CourseWork\Exception.h \
- C:\Users\NHTHS\Desktop\CourseWork\SyntaxChecker.h \
- C:\Users\NHTHS\Desktop\CourseWork\ConfigMapper.h
+ C:\Users\NHTHS\Desktop\CourseWork\Exception.h
