+++ cmake-build-debug/CMakeFiles/Makefile2
@@ -80,12 +80,12 @@ clean: CMakeFiles/CourseWork.dir/clean
 CMakeFiles/CourseWork.dir/all:
 	$(MAKE) $(MAKESILENT) -f CMakeFiles\CourseWork.dir\build.make CMakeFiles/CourseWork.dir/depend
 	$(MAKE) $(MAKESILENT) -f CMakeFiles\CourseWork.dir\build.make CMakeFiles/CourseWork.dir/build
-	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\NHTHS\Desktop\CourseWork\cmake-build-debug\CMakeFiles --progress-num=1,2,3,4,5,6,7,8,9,10,11 "Built target CourseWork"
+	@$(CMAKE_COMMAND) -E cmake_echo_color --switch=$(COLOR) --progress-dir=C:\Users\NHTHS\Desktop\CourseWork\cmake-build-debug\CMakeFiles --progress-num=1,2,3,4,5,6,7,8,9,10 "Built target CourseWork"
 .PHONY : CMakeFiles/CourseWork.dir/all
 
 # Build rule for subdir invocation for target.
 CMakeFiles/CourseWork.dir/rule: cmake_check_build_system
-	$(CMAKE_COMMAND) -E cmake_progress_start C:\Users\NHTHS\Desktop\CourseWork\cmake-build-debug\CMakeFiles 11
+	$(CMAKE_COMMAND) -E cmake_progress_start C:\Users\NHTHS\Desktop\CourseWork\cmake-build-debug\CMakeFiles 10
 	$(MAKE) $(MAKESILENT) -f CMakeFiles\Makefile2 CMakeFiles/CourseWork.dir/all
 	$(CMAKE_COMMAND) -E cmake_progress_start C:\Users\NHTHS\Desktop\CourseWork\cmake-build-debug\CMakeFiles 0
 .PHONY : CMakeFiles/CourseWork.dir/rule
