+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Nov 26 01:53 RTZ 2 (ceia)
+Start testing: Nov 28 13:18 RTZ 2 (ceia)
 ----------------------------------------------------------
-End testing: Nov 26 01:53 RTZ 2 (ceia)
+End testing: Nov 28 13:18 RTZ 2 (ceia)
