+++ main.cpp
@@ -1,13 +1,12 @@
-#include "ArciverExecutor.h"
+#include "ArchiverExecutor.h"
 #include <iostream>
 #include <string>
 
-#include "Exception.h"
-#include "SyntaxChecker.h"
+#include "utils/Exception.h"
 
-#include "ConfigMapper.h"
+#include "utils/ConfigMapper.h"
 
-#include "Log.h"
+#include "utils/Log.h"
 
 using namespace std;
 
@@ -18,7 +17,10 @@ int main(int argc, char *argv[]) {
         ArchiverExecutor::execute(config);
     } catch (Exception &e) {
         cerr << e.what() << endl;
-        cerr << SyntaxChecker::printErr() << endl;
+        cerr << "Wrong arguments\n"
+                "Use: algorithmFano.exe [mode] [view] [path]\n"
+                "Example 1: algorithmFano.exe -c -v c:/file.txt\n"
+                "Example 2: algorithmFano.exe -d c:/file.txt.archive" << endl;
         return -1;
     }
